+++ src/main/java/edu/kit/ipd/crowdcontrol/objectservice/duplicateDetection/Similarity/StringSimilarity.java
@@ -1,11 +1,10 @@
-package edu.kit.ipd.crowdcontrol.objectservice.duplicateDetection;
+package edu.kit.ipd.crowdcontrol.objectservice.duplicateDetection.Similarity;
 
 import com.google.common.collect.Sets;
 
 import java.nio.ByteBuffer;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
-import java.util.List;
 import java.util.Set;
 
 /**
@@ -37,21 +36,11 @@ public class StringSimilarity {
         long hash1 = computeSimhashFromShingles(shingle1);
         long hash2 = computeSimhashFromShingles(shingle2);
 
-        return getHammingDistanceOfHashes(hash1,hash2);
+        return HashSimilarity.getHammingDistanceOfHashes(hash1,hash2);
     }
 
 
-    /**
-     * Calculates the hamming-distance of two given hashes
-     * @param hash1 first hash
-     * @param hash2 second hash
-     * @return hamming distance of the specified hashes
-     */
-    public static int getHammingDistanceOfHashes(long hash1, long hash2){
-        long xor = hash1 ^ hash2;
-        return Long.bitCount(xor);
 
-    }
 
     /**
      * Computes the jaccard-coefficient of two stings.
@@ -83,24 +72,12 @@ public class StringSimilarity {
         long hash1 = computeSimhashFromShingles(shingle1);
         long hash2 = computeSimhashFromShingles(shingle2);
 
-        return getSimilarityFromHash(hash1,hash2);
+        return HashSimilarity.getSimilarityFromHash(hash1,hash2);
     }
 
 
 
-    /**
-     * Computes the similarity of two hashes via simhash
-     * @param hash1 first hash
-     * @param hash2 second hash
-     * @return the similarity of the two given hashes
-     */
-    public static double getSimilarityFromHash(long hash1, long hash2){
-        long xor = hash1 ^ hash2;
-        if(xor == 0){
-            return 1;
-        }
-        return 1.0 - (((double) Long.bitCount(xor) + 1) / (65 - Long.numberOfLeadingZeros(xor)));
-    }
+
 
     /**
      * Computes hashing of the given set of shingles
